requirements.txt
pillow>=9.0.0
opencv-python>=4.6.0
numpy>=1.21.0
beautifulsoup4>=4.11.0
requests>=2.28.0
pandas>=1.4.0
openpyxl>=3.0.0